Don't rebuild the entire frame for each client
We can reuse most of the work that went into building the frame except for
timestamp information, which we have to patch in individually for each client.
Doing this lets us avoid duplicating the rest of the frame as well as use only
one buffer for sending to as many HTTP clients as we have.